Include reference to UDTT in opening paragraph
Discussion within this StackOverflow thread made it evident that there was some ambiguity about the need for CLR to be enabled to allow creation of user-defined table types.  Some investigation of SQL 2005 Books Online shows it apparent that the opening paragraph of this particular docs page has not been modified since that version, to account for additional "user-defined" types (particularly TABLE) that are available in later versions.